Miscellaneous 64-bit and other issues  [Oliver Stoeneberg]
git-svn-id: svn://messdev.no-ip.org/mess@755 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/tools/messtest/core.c
@@ -130,7 +130,7 @@ void messtest_get_data(xml_data_node *node, mess_pile *pile)
 				else if (c == '*')
 				{
 					blobstate = BLOBSTATE_AFTER_STAR;
-					multiple = (size_t) -1;
+					multiple = -1;
 				}
 				else if (c == '\'')
 				{
@@ -157,13 +157,13 @@ void messtest_get_data(xml_data_node *node, mess_pile *pile)
 				if (isdigit(c))
 				{
 					/* add this digit to the multiple */
-					if (multiple == (size_t) -1)
+					if (multiple == -1)
 						multiple = 0;
 					else
 						multiple *= 10;
 					multiple += c - '0';
 				}
-				else if ((c != '\0') && isspace(c) && (multiple == (size_t) -1))
+				else if ((c != '\0') && isspace(c) && (multiple == -1))
 				{
 					/* ignore whitespace */
 				}
